+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version latest (0.15.1rc?, 2019-09-24?)
 
 - ...
 - enh: simplify cli.cost code by gathering all the mainManager setup code into `pipeline_factory` files
+- enh: refactor to rename `*Listener` to `Calculator*Ec2` and `Analyzer*` to `Calculator*Redshift`
 
 
 Version 0.15.1 (2019-11-20)